+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 5.3.0 (2023-10-27)
+### Features Added
+
+- New struct `ResourceSharingProfile`
+- New field `SharedSubscriptionIDs` in struct `CapacityReservationGroupInstanceView`
+- New field `SharingProfile` in struct `CapacityReservationGroupProperties`
+
+
 ## 5.3.0-beta.1 (2023-10-09)
 ### Features Added